Perform discovery API calls only as needed
This commit changes the discovery REST mapper that was used to be targeted to
that specific resource version.

This is to reduce load on the Kubernetes API as well as to reduce the chance
of long delays due to queries per second (QPS) exceeding the default limit.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 670631dc0c227a6d8ec5ecf5b5d66bf9c77752b1)