fix nil reference when hub-only controllers are disabled
When --disable-spec-sync is set to true, the hub manager is set to nil.
Currently, this is not detected when setting up the controllers, which
causes a seg fault when referencing the nil hub manager. In addition to
fixing the behaviour, a new cache will be created on the hub for the case
where the hub manager, and subsequently, the related hub manager cache,
is not set.

ref: https://issues.redhat.com/browse/ACM-6869
Signed-off-by: Jason Zhang <jaszhang@redhat.com>
(cherry picked from commit 0ac24d02c82226e34d8bd05933ce00e6a63b02c2)