Set template-error on correct details item
When adding a template-error to the status for a template that can't be
decoded, the status-sync was incorrectly appending to the details array
twice. So for something that would be called "template-0", the 0th item
in the details array would be almost empty, and the 1st item would have
all of the status events. This was causing some events to be emitted
repeatedly, since the template-sync logic to prevent this was looking at
the 0th item, which would never get the event.

This commit also cleans up some awkward pointer gymnastics.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit c009fa45685dce4c9091adc2793dccfaedf4ae74)