Report Pending status despite hub template error
A previous change allowing hub-templates for all policy types, and
handling any hub-template errors in this controller, caused a change in
behavior. Before that change, when a policy both had a hub-template
error *and* an unsatisfied dependency, the policy would be marked as
Pending. The change made the hub-template error have priority.

This commit gives the Pending status priority, and adds a test for that
specific behavior. It also handles ConfigurationPolicies particularly
carefully: setting `PruneObjectBehavior` to `None` when the template
will be deleted specifically because of a hub-template error.

Refs:
 - https://issues.redhat.com/browse/ACM-11530

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit d780b0e587eadf6abc9937486a395aab38a740f3)