Experiment 1
Trying running some in-repo tests, it didn't seem easy to setup KinD in
a container that could be run in the pipeline, so we'll not worry about
that right now. This just does some lint-ish kinds of checks.

This also removes some `runAfter` bits that I don't *think* are
necessary, I'm curious if it still works, and if the renovate bot puts
them back, or what.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>