Trigger status-sync reconciles when spec-sync sees mismatching status
Since status-sync watches the managed cluster's replicated policies on
the managed cluster (not the hub), it has a blind spot for when the
hub's status doesn't match and nothing triggers a status-sync reconcile
(e.g. new compliance event or policy update). Since spec-sync is already
watching the managed cluster's replicated policies on the hub, it can do
the work of detecting the status mismatch on the hub and managed
cluster, and trigger a status-sync reconcile to have the status-sync
controller fix it.

Relates:
https://issues.redhat.com/browse/ACM-12447

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit de1345afe0a6308a355082e5204613fbe08d43e0)