Deduplicate objects to watch
This version of the dependency-watches library can panic if passed
duplicate object IDs along with an object ID for a kind that is not
present on the cluster. This avoids triggering that situation, and
might be easier to backport than changes to the library itself.

Refs:
 - https://issues.redhat.com/browse/ACM-14091

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 06186f4003f6c1109d23779d43dc2935cdaf2be2)